+++ packages/isolate_bloc/lib/src/common/isolate/isolate_factory/i_isolate_messenger.dart
@@ -1,14 +1,12 @@
 import 'package:isolate_bloc/src/common/isolate/isolate_event.dart';
 
-typedef SendToIsolate = void Function(IsolateEvent message);
-
 /// {@template isolate_messenger}
-/// Messenger for communication between [Isolate]s.
+/// Messenger for communication between `Isolate`s.
 /// {@endtemplate}
 abstract class IIsolateMessenger {
-  /// Stream with messages from [Isolate]
+  /// Stream with messages from `Isolate`.
   Stream<IsolateEvent> get messagesStream;
 
-  /// Sends messages to the [Isolate]
+  /// Sends messages to the `Isolate`.
   void send(IsolateEvent message);
 }