+++ packages/isolate_bloc/test/blocs/error_cubit.dart
@@ -1,0 +1,12 @@
+import 'package:isolate_bloc/isolate_bloc.dart';
+
+class ErrorCubit extends IsolateCubit {
+  ErrorCubit() : super(Object());
+
+  @override
+  void onEventReceived(event) {
+    throw ExpectedException();
+  }
+}
+
+class ExpectedException {}