+++ packages/isolate_bloc/test/mock/mock_isolate_bloc_wrapper.dart
@@ -1,0 +1,4 @@
+import 'package:isolate_bloc/src/common/bloc/isolate_bloc_wrapper.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockIsolateBlocWrapper<T> extends Mock implements IsolateBlocWrapper<T> {}