+++ packages/isolate_bloc/test/test_utils/errors_matcher.dart
@@ -1,0 +1,12 @@
+import 'dart:async';
+
+import 'package:flutter_test/flutter_test.dart';
+
+T? expectError<T>(T Function() errorAction, Matcher errorMatcher) {
+  return runZonedGuarded(
+    errorAction,
+    (error, stack) {
+      expect(error, errorMatcher);
+    },
+  );
+}